Common label component for variables and layers controls
