+++ packages/scenes/src/variables/components/VariableValueSelectors.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
 
-import { GrafanaTheme2, VariableHide } from '@grafana/data';
-import { selectors } from '@grafana/e2e-selectors';
-import { Tooltip, useStyles2 } from '@grafana/ui';
+import { VariableHide } from '@grafana/data';
 
 import { SceneObjectBase } from '../../core/SceneObjectBase';
 import { sceneGraph } from '../../core/sceneGraph';
 import { SceneComponentProps, SceneObject, SceneObjectState } from '../../core/types';
 import { SceneVariableState } from '../types';
-import { css } from '@emotion/css';
+import { ControlsLabel } from '../../utils/ControlsLabel';
 
 export class VariableValueSelectors extends SceneObjectBase<SceneObjectState> {
   public static Component = VariableValueSelectorsRenderer;
@@ -16,24 +14,17 @@ export class VariableValueSelectors extends SceneObjectBase<SceneObjectState> {
 
 function VariableValueSelectorsRenderer({ model }: SceneComponentProps<VariableValueSelectors>) {
   const variables = sceneGraph.getVariables(model)!.useState();
-  const styles = useStyles2(getStyles);
 
   return (
     <>
       {variables.variables.map((variable) => (
-        <VariableValueSelectWrapper key={variable.state.key} variable={variable} styles={styles} />
+        <VariableValueSelectWrapper key={variable.state.key} variable={variable} />
       ))}
     </>
   );
 }
 
-function VariableValueSelectWrapper({
-  variable,
-  styles,
-}: {
-  variable: SceneObject<SceneVariableState>;
-  styles: VariableLabelStyles;
-}) {
+function VariableValueSelectWrapper({ variable }: { variable: SceneObject<SceneVariableState> }) {
   const state = variable.useState();
 
   if (state.hide === VariableHide.hideVariable) {
@@ -41,14 +32,18 @@ function VariableValueSelectWrapper({
   }
 
   return (
-    <div className={styles.container}>
-      <VariableLabel state={state} styles={styles} />
+    <div
+      style={{
+        display: 'flex',
+      }}
+    >
+      <VariableLabel state={state} />
       <variable.Component model={variable} />
     </div>
   );
 }
 
-function VariableLabel({ state, styles }: { state: SceneVariableState; styles: VariableLabelStyles }) {
+function VariableLabel({ state }: { state: SceneVariableState }) {
   if (state.hide === VariableHide.hideLabel) {
     return null;
   }
@@ -56,51 +51,5 @@ function VariableLabel({ state, styles }: { state: SceneVariableState; styles: V
   const elementId = `var-${state.key}`;
   const labelOrName = state.label ?? state.name;
 
-  if (state.description) {
-    return (
-      <Tooltip content={state.description} placement={'bottom'}>
-        <label
-          className={styles.variableLabel}
-          data-testid={selectors.pages.Dashboard.SubMenu.submenuItemLabels(labelOrName)}
-          htmlFor={elementId}
-        >
-          {labelOrName}
-        </label>
-      </Tooltip>
-    );
-  }
-
-  return (
-    <label
-      className={styles.variableLabel}
-      data-testid={selectors.pages.Dashboard.SubMenu.submenuItemLabels(labelOrName)}
-      htmlFor={elementId}
-    >
-      {labelOrName}
-    </label>
-  );
+  return <ControlsLabel htmlFor={elementId} label={labelOrName} description={state.description ?? undefined} />;
 }
-
-const getStyles = (theme: GrafanaTheme2) => ({
-  container: css({
-    display: 'flex',
-  }),
-  variableLabel: css({
-    background: theme.isDark ? theme.colors.background.primary : theme.colors.background.secondary,
-    display: `flex`,
-    alignItems: 'center',
-    padding: theme.spacing(0, 1),
-    fontWeight: theme.typography.fontWeightMedium,
-    fontSize: theme.typography.bodySmall.fontSize,
-    height: theme.spacing(theme.components.height.md),
-    lineHeight: theme.spacing(theme.components.height.md),
-    borderRadius: theme.shape.borderRadius(1),
-    border: `1px solid ${theme.components.input.borderColor}`,
-    position: 'relative',
-    // To make the border line up with the input border
-    right: -1,
-    whiteSpace: 'nowrap',
-  }),
-});
-
-type VariableLabelStyles = ReturnType<typeof getStyles>;
